Fix possible errors related to input peer resolving in documents, implement context storing and versioning for FileRefId
And also implement media spec and caching for polls

+++ mtproto/src/main/java/telegram4j/mtproto/MTProtoRetrySpec.java
@@ -40,7 +40,7 @@ public class MTProtoRetrySpec extends Retry {
                     rpcExc.getError().errorMessage().startsWith("FLOOD_WAIT_")) {
                 String arg = rpcExc.getError().errorMessage().substring(
                         rpcExc.getError().errorMessage().lastIndexOf('_') + 1);
-                Duration delay = Duration.ofSeconds(Long.parseLong(arg));
+                Duration delay = Duration.ofSeconds(Integer.parseInt(arg));
                 return Mono.delay(delay);
             }
             return Mono.error(currentFailure);
